move: Move config package
