+++ p2pCraft-api/src/main/java/io/github/seujorgenochurras/p2pApi/api/config/CorsConfig.java
@@ -1,11 +1,11 @@
-package io.github.seujorgenochurras.p2pApi.common;
+package io.github.seujorgenochurras.p2pApi.api.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
-public class Config implements WebMvcConfigurer {
+public class CorsConfig implements WebMvcConfigurer {
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
